Use LENGTH_PREFIX_SIZE constant everywhere
